Issue-15 Separate KafkaSerde into config class, add schema-registry commands, add --detail parameter to query to get offsets and partitions, help text on all commands, reasonable connection timeout on startup, logical type object substitution for avro

+++ query-shell/src/main/java/org/galatea/kafka/shell/domain/DbRecordKey.java
@@ -5,7 +5,7 @@ import lombok.Value;
 @Value
 public class DbRecordKey {
 
-  private MutableField<Long> partition = new MutableField<>(-1L);
+  private MutableField<Integer> partition = new MutableField<>(-1);
   private MutableField<Long> offset = new MutableField<>(-1L);
   private MutableField<String> stringKey = new MutableField<>(null);
 }
